+++ deployment/metrics/variables.tf
@@ -17,5 +17,8 @@
 variable "project" {
   type     = string
   nullable = false
-  default  = ""
+  validation {
+    condition     = var.project != ""
+    error_message = "project is required."
+  }
 }